Fix more shadowed variables; Fix assert failure on exit due to NPC gfx

+++ src/NPC.cpp
@@ -321,23 +321,23 @@ void NPC::logic() {
 	{
 		if (mapr->events[i].npcName == name)
 		{
-			mapr->events[i].location.x = stats.pos.x;
-			mapr->events[i].location.y = stats.pos.y;
+			mapr->events[i].location.x = static_cast<int>(stats.pos.x);
+			mapr->events[i].location.y = static_cast<int>(stats.pos.y);
 
-			mapr->events[i].hotspot.x = stats.pos.x;
-			mapr->events[i].hotspot.y = stats.pos.y;
+			mapr->events[i].hotspot.x = static_cast<int>(stats.pos.x);
+			mapr->events[i].hotspot.y = static_cast<int>(stats.pos.y);
 
 			mapr->events[i].center.x =
 				static_cast<float>(stats.pos.x) + static_cast<float>(mapr->events[i].hotspot.w)/2;
 			mapr->events[i].center.y =
 				static_cast<float>(stats.pos.y) + static_cast<float>(mapr->events[i].hotspot.h)/2;
 
-			for (int ci = 0; ci < mapr->events[i].components.size(); ci++)
+			for (size_t ci = 0; ci < mapr->events[i].components.size(); ci++)
 			{
 				if (mapr->events[i].components[ci].type == EC_NPC_HOTSPOT)
 				{
-					mapr->events[i].components[ci].x = stats.pos.x;
-					mapr->events[i].components[ci].y = stats.pos.y;
+					mapr->events[i].components[ci].x = static_cast<int>(stats.pos.x);
+					mapr->events[i].components[ci].y = static_cast<int>(stats.pos.y);
 				}
 			}
 		}
